Fix user resource authorization
Routing on auth error did work, but we didn't return from controller
methods which got executed resulting in unexpected exceptions. Throw
custom exceptions instead when user is not authorized and handle them in
the application controller.

Rename "authenticate" to "authorizeRole ", as it was authorizing a
resource based on a role in the end. Authentication is done by the
methods provided by "has_secure_password" [1].

Move "authorizeRole" to User, as this should be a responsibility of the
user and not of the application controller (analogically "authenticate"
is a method of User [1]).

[1]: http://api.rubyonrails.org/classes/ActiveModel/SecurePassword/ClassMethods.html#method-i-has_secure_password

Change-Id: Iac92b99a4f7eaeeb9d8d01594508f77ec55481ed